chore: remove unnecessary Arc wrappers around DB
A fairly common pattern in types which implement clients (db, web, etc.)
is to have an inner type which is Arc wrapped. Both LanceDB and Redb are
such types. This commit replaces the unnecessary Arc wrappers around
these types in the storage module and all the places where they are used
with a simple owned type which is cloned when needed.