refactor: Index behind an interface (#462)
Lays the ground for dependency injecting the index into the application.
Instead of gross boundary violations everywhere, the command handler is
now started with an Index (DuckdbIndex here).

Exact interface and changes will likely still change, but this puts it
nicely at the seams. There's still some spots where the index could be
used but isn't.

It enables the following future changes (where as I see it, the hard
part is now done):

- Running the kwaak backend with a different database
- Running the same kwaak backend with different pipelines
- Running kwaak on multiple repositories
- Running the index in a hosted environment


Requires bosun-ai/swiftide#720 as I wanted the pipelines to be generic
over the search strategy but tools didn't support generics.